Cleaned up, fixed memleaks.
