+++ src/fs.h
@@ -10,7 +10,7 @@
 // size, and actual data.
 struct file {
     char name[FILENAME_SIZE];
-    char *bytes;
+    uint8_t *bytes;
     size_t s;
 };
